* bye bye c++11, hello c++14  * minor fixes (forgot to add the assumed lua_path in the log msg if it fails to find it; signed/unsigned conversion fix)  * updated premake file and xcode project  * updated README according to latest changes  * version bump to v0.5.1-d3

+++ src/lua_unibot.hpp
@@ -100,7 +100,7 @@ protected:
 				}
 				else {
 					script.seekg(0, ios::beg);
-					char* script_data = new char[script_size+1];
+					char* script_data = new char[(size_t)script_size + 1];
 					script.read(script_data, script_size);
 					script_data[script_size] = 0;
